Large refactor of testing scripts
