+++ testing/configs/gnb.yaml
@@ -1,13 +1,13 @@
-mcc: '910'          # Mobile Country Code value
-mnc: '54'           # Mobile Network Code value (2 or 3 digits)
+mcc: '901'          # Mobile Country Code value
+mnc: '70'           # Mobile Network Code value (2 or 3 digits)
 
 nci: '0x000000010'  # NR Cell Identity (36-bit)
 idLength: 32        # NR gNB ID length in bits [22...32]
 tac: 1              # Tracking Area Code
 
-linkIp: 192.168.60.200   # gNB's local IP address for Radio Link Simulation (Usually same with local IP)
-ngapIp: 192.168.60.200   # gNB's local IP address for N2 Interface (Usually same with local IP)
-gtpIp: 192.168.60.200    # gNB's local IP address for N3 Interface (Usually same with local IP)
+linkIp: 192.168.56.200   # gNB's local IP address for Radio Link Simulation (Usually same with local IP)
+ngapIp: 192.168.56.200   # gNB's local IP address for N2 Interface (Usually same with local IP)
+gtpIp: 192.168.56.200    # gNB's local IP address for N3 Interface (Usually same with local IP)
 
 # List of AMF address information
 amfConfigs:
