+++ testing/perf/dauth_testing/sample/basic/home_setup.yaml
