+++ testing/perf/dauth_testing/sample/scripts/full_network_tests.sh
