+++ testing/perf/metrics.py
@@ -105,7 +105,7 @@ class PerfMetrics:
             '{
                 "nanoseconds_since_auth":<int>,
                 "nanoseconds_since_registration":<int>,
-                "ue_supi":"91054xxxxxxxxxx",
+                "ue_supi":"90170xxxxxxxxxx",
              }'
         """
         dataset = json.loads(json_string)
