+++ testing/perf/sample_configs/vagrant.yaml
@@ -1,0 +1,35 @@
+# This is an example of a full vagrant VM network
+
+directory:
+  hostname: "directory"
+  id: "directory"
+  is_vagrant: true
+  directory_addr: 192.168.56.250:8900  # vagrant only
+
+ueransim:
+  hostname: "ueransim"
+  id: "ueransim"
+  is_vagrant: true
+
+services:
+  # First service is always the home network
+  - hostname: "colte1"
+    id: "colte-1"
+    is_vagrant: true
+    service_ip: "192.168.56.101"  # vagrant only
+
+  # Second service is always the serving network (when applicable)
+  - hostname: "colte2"
+    id: "colte-2"
+    is_vagrant: true
+    service_ip: "192.168.56.102"  # vagrant only
+
+  # Remaining services are backups (up to 31 services)
+  - hostname: "colte3"
+    id: "colte-3"
+    is_vagrant: true
+    service_ip: "192.168.56.103"  # vagrant only
+  - hostname: "colte4"
+    id: "colte-4"
+    is_vagrant: true
+    service_ip: "192.168.56.104"  # vagrant only
