--- testing/pyproject.toml
@@ -1,17 +1,0 @@
-[tool.poetry]
-name = "dauth-testing"
-version = "0.0.1"
-description = "Testing scripts for dauth"
-authors = ["Nick Durand <durandn@cs.washington.edu>"]
-
-[tool.poetry.dependencies]
-python = "^3.8"
-setuptools = "^65.5.1"
-paramiko = "^2.11.0"
-PyYAML = "^6.0"
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
